\name{PASolve}
\alias{PASolve}
\title{Function to submit a job to the ProActive Scheduler}
\description{
  This function executes a method of a Java class
  defined in a JAR inside the R package.
}
\usage{
  PASolve("function_name", args_vector)
}
\details{
  Details are given in the package vignette.
}
\value{
  The string "Hello Java World!" is returned.
}
\author{Tobias Verbeke}
\examples{
}
\keyword{manip}