\name{PADebug}
\alias{PADebug}
\title{sets PARConnector Debug mode}
\usage{
PADebug(debug = FALSE)
}
\arguments{
  \item{debug}{to set the debug mode to on (TRUE) or off
  (FALSE)}
}
\value{
the current or new state of the debug mode
}
\description{
PADebug can be used either to set the Debug mode to on/off
or to know the current state of the debug mode.
}
\details{
In Debug mode a lot of verbose information will be printed
(detailed content of PATask created, code analysis
debugging, remote execution trace)
}

